Adds support for including DLL files
Allows including optional DLL files in the installer.

The changes introduce a new `DllEntry` model to represent DLL file information,
and updates the configuration to include a list of `DllEntry` objects.

The script builder is updated to copy these DLLs to the application directory
during installation. This avoids the need for users to install MS Visual C++ redistributable.