Adds option to download VC++ redistributable
Implements a feature to allow users to download the Visual C++ Redistributable package directly from the installer.

This adds a checkbox to the installer's final page that, when checked, opens the user's browser to download the VC++ redistributable after installation is complete.

This feature is enabled by setting `vc_redist: download` in the `pubspec.yaml` file.